refactor: index.tsx 파일 없는 폴더는 index.styles.tsx 파일 삭제
