+++ src/components/Book/Budget/createBudgetModal.tsx
@@ -1,12 +1,78 @@
 import React from "react";
+import styled from "styled-components";
 import { useRecoilState } from "recoil";
 import { useParams } from "react-router-dom";
-import * as S from "./index.styles";
-import COLORS from "../../../constants/color";
 import { createBudgetModalAtom } from "../../../stores/atoms/context";
 import useBudget from "../../../hooks/book/useBudget";
 import useCreateBudget from "../../../hooks/book/useCreateBudget";
 import getMoneyUnit from "../../../utils/money";
+import COLORS from "../../../constants/color";
+import Button from "../../Button";
+import Input from "../../Input";
+
+const BudgetContainer = styled.div`
+  width: 41.2rem;
+  position: relative;
+  padding: 3rem 0;
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+  h2 {
+    font-weight: 700;
+    font-size: 2.3rem;
+  }
+  p {
+    margin-top: 1rem;
+    font-size: 1.3rem;
+    color: ${COLORS.GREY[400]};
+  }
+`;
+const CreateButton = styled(Button)`
+  font-size: 1.4rem;
+  width: 100%;
+  border-radius: 0.5rem;
+  padding: 1.3rem 0;
+  font-weight: 700;
+  margin-top: 7rem;
+`;
+
+const BudgetInput = styled(Input)`
+  margin-top: 1rem;
+  width: 100%;
+  border-radius: 0.5rem;
+  input {
+    padding-right: 3rem;
+  }
+`;
+const LastMonthInfo = styled.div`
+  background-color: ${COLORS.LIGHT_BLUE};
+  border-radius: 0.5rem;
+  padding: 0 1.2rem 1.2rem 1.2rem;
+  width: 14rem;
+  position: absolute;
+  right: 0;
+  display: flex;
+  flex-direction: column;
+  top: 20.2rem;
+  div {
+    display: flex;
+    gap: 10px;
+  }
+  p {
+    font-size: 1rem;
+  }
+`;
+
+const Won = styled.div`
+  position: absolute;
+  bottom: 1.5rem;
+  right: 1rem;
+  font-size: 1.4rem;
+`;
+
+const InputWrapper = styled.div`
+  position: relative;
+`;
 
 export default function Create() {
   const [, setCreateModalOpen] = useRecoilState(createBudgetModalAtom);
@@ -40,13 +106,13 @@ export default function Create() {
     }
   };
   return (
-    <S.BudgetContainer>
+    <BudgetContainer>
       <div>
         <h2>{date.getMonth() + 1}월 예산 설정</h2>
         <p>이번 달도 잘 해내실거라고 생각해요!&nbsp;저희가 응원합니다 💪🏻</p>
       </div>
-      <S.InputWrapper>
-        <S.BudgetInput
+      <InputWrapper>
+        <BudgetInput
           size={38}
           value={budgets}
           label="금액"
@@ -56,9 +122,9 @@ export default function Create() {
           }
           onChange={setBudget}
         />
-        <S.Won>원</S.Won>
-      </S.InputWrapper>
-      <S.LastMonthInfo>
+        <Won>원</Won>
+      </InputWrapper>
+      <LastMonthInfo>
         <div>
           <p>지난 달 예산</p>
           <p style={{ color: COLORS.BLUE }}>
@@ -74,8 +140,8 @@ export default function Create() {
             <p style={{ color: COLORS.YELLOW }}>초과</p>
           )}
         </div>
-      </S.LastMonthInfo>
-      <S.CreateButton
+      </LastMonthInfo>
+      <CreateButton
         onClick={() => {
           selectUpdateCreate();
           setCreateModalOpen(false);
@@ -83,6 +149,6 @@ export default function Create() {
         width="100%"
         title={thisMonthData.budget.value ? "예산 수정" : "예산 등록"}
       />
-    </S.BudgetContainer>
+    </BudgetContainer>
   );
 }
